Add Design docs to provide provide overview and do some renaming to make more consistent
