+++ src/fake-schema/field/directives/example/example.test.ts
@@ -1,4 +1,4 @@
-import { resolveExample, resolveValues } from "./example";
+import { resolveExample } from ".";
 
 // re-align `typeFieldMap` and `fieldMap` (resolve examples and fakes), using a generic `resultResolver`.
 // Allow `matches` list for both, using `resolveMatches`
@@ -17,6 +17,8 @@ describe("resolveExample", () => {
     });
   });
 });
+
+import { resolveValues } from "./ExampleValue";
 
 describe("resolveValues", () => {
   const obj = {
