+++ src/fake-schema/field/directives/fake/FakeValue.ts
@@ -1,0 +1,1 @@
+export class FakeValue {}
