+++ src/fake-schema/field/directives/fake/fakers/resolve/index.ts
@@ -1,3 +1,3 @@
-export { resolveExample } from "../../../example/example";
+export { resolveExample } from "../../../example";
 export { resolveFake } from "./fake";
 export { error } from "./error";
