+++ src/fake-schema/field/directives/fake/index.ts
@@ -1,1 +1,3 @@
 // TODO: move fakeValue here and turn into a class
+import * as fakers from "./fakers";
+export { fakers };
